Fix issue hide/show race by cancelling the final hide callback on show
Added a _completeHideTimer to track the final run.later which actually completes the hiding of the tooltip. This is different than the _hideTimer which tracks a desired hide of the tooltip based on a duration supplied by the user.

This timer gets cancelled on every show to prevent a race condition between showing and hiding the tooltip.